requirements.txt
numpy
scipy>=1.0.0
scikit-learn
mne
pyyaml
matplotlib==3.0.3
psychopy
pyglet==1.3.1
pyedflib
keyboard
joblib
seaborn
